requirements.txt
certifi==2020.12.5
chardet==4.0.0
fire==0.4.0
idna==2.10
python-dotenv==0.16.0
requests==2.25.1
six==1.16.0
termcolor==1.1.0
urllib3==1.26.4
